requirements.txt
requests>=2.31.0
openai>=1.30.0
python-dotenv>=1.0.0
Jinja2>=3.1.2
pytest>=8.0.0
pytest-mock>=3.12.0
